# Faster Generic Hashtable.