--- a/src/hashtable.c
+++ b/src/hashtable.c
@@ -23,15 +23,32 @@
 #include <stdio.h>
 #include "hashtable.h"
 
+/* Open addressing works best if it can take advantage of memory caches using
+ * locality for probes of adjacent buckets on collisions. So we pack the keys
+ * tightly together in their own key table and avoid referencing the element
+ * table and elements as much as possible. Key value zero is reserved as a
+ * marker for an empty bucket to avoid checking for NULL in the element table.
+ * If we do get a hash value of zero, we -1 to wrap it around to 0xffff. */
+
+/* Use max 0.8 load factor to avoid bad open addressing performance. */
+#define HASHTABLE_LOADFACTOR_NUM 8
+#define HASHTABLE_LOADFACTOR_DEN 10
+
 hashtable_t *hashtable_new(int size, hash_f hash, cmp_f cmp)
 {
     hashtable_t *t;
     int size2;
 
-    /* Use next power of 2 larger than 2x the requested size. */
-    for (size2 = 1; size2 < 2 * size; size2 <<= 1) ;
-    if (!(t = calloc(1, sizeof(hashtable_t)+ size2 * sizeof(void *))))
+    /* Adjust requested size to account for max load factor. */
+    size = 1 + size * HASHTABLE_LOADFACTOR_DEN / HASHTABLE_LOADFACTOR_NUM;
+    /* Use next power of 2 larger than the requested size. */
+    for (size2 = 1; size2 < size; size2 <<= 1) ;
+    if (!(t = calloc(1, sizeof(hashtable_t)+ size2 * sizeof(unsigned))))
         return NULL;
+    if (!(t->etable = calloc(size2, sizeof(void *)))) {
+        free(t);
+        return NULL;
+    }
     t->size = size2;
     t->count = 0;
     t->hash = hash;
@@ -41,12 +58,15 @@ hashtable_t *hashtable_new(int size, hash_f hash, cmp_f cmp)
 
 void hashtable_free(hashtable_t *t)
 {
-    free(t);
+    if (t) {
+        free(t->etable);
+        free(t);
+    }
 }
 
+/* MurmurHash3 finalization mix function. */
 static inline unsigned mix32(unsigned int h)
 {
-    /* MurmurHash3 finalization mix function. */
     h ^= h >> 16;
     h *= 0x85ebca6b;
     h ^= h >> 13;
@@ -55,10 +75,17 @@ static inline unsigned mix32(unsigned int h)
     return h;
 }
 
+/* Get hash key, reserving zero as an empty bucket marker. */
+static inline unsigned get_key(hashtable_t *t, void *e)
+{
+    unsigned k = t->hash(e);
+    return k ? k : k - 1;
+}
+
 /* Prefix macro for probing table t for key k with index i. */
 #define do_probe(t, k) \
     unsigned mask = t->size - 1;\
-    unsigned index = mix32(t->hash(k)) & mask;\
+    unsigned index = mix32(k) & mask;\
     unsigned i = index, s = 0;\
     do
 
@@ -69,22 +96,28 @@ static inline unsigned mix32(unsigned int h)
 void *hashtable_add(hashtable_t *t, void *e)
 {
     assert(e != NULL);
-    do_probe(t, e) {
-        if (!t->table[i]) {
+    unsigned k = get_key(t, e);
+
+    do_probe(t, k) {
+        if (!t->ktable[i]) {
             t->count++;
-            return t->table[i] = e;
+            t->ktable[i] = k;
+            return t->etable[i] = e;
         }
     } while_probe;
     return NULL;
 }
 
-void *hashtable_find(hashtable_t *t, void *k)
+void *hashtable_find(hashtable_t *t, void *m)
 {
-    assert(k != NULL);
+    assert(m != NULL);
+    unsigned km = get_key(t, m), ke;
     void *e;
 
-    do_probe(t, k) {
-        if (!(e = t->table[i]) || !t->cmp(k, e))
+    do_probe(t, km) {
+        if (!(ke = t->ktable[i]))
+            return NULL;
+        if (km == ke && !t->cmp(m, e = t->etable[i]))
             return e;
     } while_probe;
     return NULL;
@@ -104,11 +137,11 @@ void *hashtable_next(hashtable_iter_t *i)
     assert(i->htable != NULL);
     assert(i->index <= i->htable->size);
     hashtable_t *t = i->htable;
+    void *e;
 
     while (i->index < t->size) {
-        if (t->table[i->index])
-            return t->table[i->index++];
-        i->index++;
+        if ((e = t->etable[i->index++]))
+            return e;
     }
     return NULL;
 }