This is an alternative implementation to https://github.com/librsync/librsync/pull/79 that is about the same speed but maybe a little cleaner implementation with more tests. It also includes optional extra match stats with --stats, reduced memory usage by packing blocksigs using --sum-size, and much faster delta calculations for large files.

On an Intel(R) Atom(TM) CPU D525 @ 1.80GHz, for a 1GiByte file with 50% different data, using a 1KiByte block-size and 8Byte sum-size for a RELEASE build, it is about 4x as fast. Comparing the current master branch;

$ time ../rdiff -f -s -I 32768 -O 32768 delta /home/abo/tmp/sig.1024K /home/abo/tmp/new.1024K /home/abo/tmp/delta.1024K
librsync: loadsig statistics: signature[1048576 blocks, 1024 bytes per block] speed[12.6 MB (12.6 MB/s) in, 0.0 MB (0.0 MB/s) out, 1 sec]
librsync: delta statistics: literal[16384 cmds, 536870912 bytes, 49152 cmdbytes] copy[2 cmds, 536870912 bytes, 0 false, 18 cmdbytes] speed[1073.7 MB (1.8 MB/s) in, 536.9 MB (0.9 MB/s) out, 593 sec]
587.20user 3.74system 9:55.86elapsed 99%CPU (0avgtext+0avgdata 51396maxresident)k
2097168inputs+1048704outputs (1major+12528minor)pagefaults 0swaps

with this branch;

$ time ../rdiff -f -s -I 32768 -O 32768 delta /home/abo/tmp/sig.1024K /home/abo/tmp/new.1024K /home/abo/tmp/delta.1024K
librsync: loadsig statistics: signature[1048576 blocks, 1024 bytes per block] speed[12.6 MB (12.6 MB/s) in, 0.0 MB (0.0 MB/s) out, 1 sec]
librsync: match statistics: signature[537395200 searches, 524288 (0.098%) matches, 627469786 (1.168x) weak sum compares, 1548850 (0.288%) strong sum compares, 1547240 (0.288%) strong sum calcs]
librsync: delta statistics: literal[16384 cmds, 536870912 bytes, 49152 cmdbytes] copy[2 cmds, 536870912 bytes, 0 false, 18 cmdbytes] speed[1073.7 MB (7.6 MB/s) in, 536.9 MB (3.8 MB/s) out, 141 sec]
131.24user 3.78system 2:24.41elapsed 93%CPU (0avgtext+0avgdata 38552maxresident)k
2097168inputs+1048696outputs (1major+11378minor)pagefaults 0swaps

Note the extra match stats output, maxresident 38552K vs 51396K, and 141sec vs 593sec. For larger files it will be even faster.