@dbaarda 2016-11-01
A quick note for anyone looking at the timing of these tests in horror;

1) I'm intentionally using a crappy CPU to highlight bottlenecks. Modern workstation CPU's run these tests many times faster than this.

2) I'm using a 1K blocksize to maximize the stress on the hashtable. This makes it many times slower than using the recommended blocksize=sqrtt(filesize).

3) The files I'm reading/writing to are on nfs mounted from a QNAP NAS, though delta calculation is CPU and memory cache bound, so this actually has little affect.