@dbaarda 2017-01-30
It's interesting you got slowdowns for smaller files. On my crappy Atom the new code was consistently faster than the old code even for small files. I suspect that on my platform even 32MiB files were stressing the L2 cache for the old hashtable, and the new hashtable has a bit better locality behavior. On a platform with more cache both old and new don't stress the cache and other factors start to dominate.

I'll have a look at some profiling and micro-optimizations to fix that after this goes in.