@dbaarda 2017-02-02
Um... it appears I have permissions to merge this pull request, which surprises me. I thought only sourcefrog could do that.

Are people OK for me to merge this, or should I leave that up to sourcefrog?