@dbaarda 2017-02-08
It's been a week since I asked if anyone would object to me merging this. If no one objects in the next day or so I'll merge it.